Vinay | #1054 | Fix build. Temp fix. Point to Bahmni test instead of webservices.rest
The original superclass somehow messes up with the regular admin/test password used for tests.
This commit bypasses the original superclass with what we have in our omod.
This will fix the build. The downside is that the superclass by itself adds some tests
which will now not be run. Investigation is still underway.